+++ todoman/model.py
@@ -544,7 +544,8 @@ class Cache:
         return rv
 
     def todos(self, all=False, lists=[], urgent=False, location='',
-              category='', grep='', sort=[], reverse=True, due=None):
+              category='', grep='', sort=[], reverse=True, due=None,
+              start=None):
         """
         Returns filtered cached todos, in a specified order.
 
@@ -567,6 +568,7 @@ class Cache:
             with a ``-`` prepended will be used to sort in reverse order.
         :param bool reverse: Reverse the order of the todos after sorting.
         :param int due: Return only todos due within ``due`` hours.
+        :param start: Return only todos before/after ``start`` date
         :return: A sorted, filtered list of todos.
         :rtype: generator
         """
@@ -603,7 +605,14 @@ class Cache:
             max_due = (datetime.now() + timedelta(hours=due)).timestamp()
             extra_where.append('AND due IS NOT NULL AND due < ?')
             params.append(max_due)
-
+        if start:
+            is_before, dt = start
+            if is_before:
+                extra_where.append('AND created_at <= ?')
+                params.append(dt)
+            else:
+                extra_where.append('AND created_at >= ?')
+                params.append(dt)
         if sort:
             order = []
             for s in sort: