+++ CHANGELOG.rst
@@ -13,6 +13,9 @@ v2.2.0
 * Use the system's date format as a default.
 * Show "Today" or "Tomorrow" when due date is today/tomorrow respectively.
 * Add list selector to TodoEditor
+* Add ``--start="[before|after] <date>`` option for ``todo list``. It shows todos
+  created before/after given date.
+
 
 v2.1.0
 ------
@@ -58,3 +61,4 @@ Packaging changes
 * New runtime dependency: configobj
 * New runtime dependency: python-dateutil
 * New test dependency: flake8-import-order.
+